requirements.txt
pyTelegramBotAPI==3.7.3
python-dotenv==0.14.0